+++ src/briefy/leica/tools/move_to_accepted.py
@@ -1,5 +1,5 @@
 """Move Order from delivered to accepted."""
-from briefy.leica.tasks.order import move_order_accepted
+from briefy.leica.tasks.order import move_orders_accepted
 from briefy.leica.db import Session
 from briefy.leica.sync import db
 
@@ -9,4 +9,4 @@ import transaction
 if __name__ == '__main__':
     db.configure(Session)
     with transaction.manager:
-        move_order_accepted()
+        move_orders_accepted()